I foget to commit.

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -3,12 +3,21 @@ const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
 const { transport, makeANiceEmail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const Mutations = {
     async createItem(parent, args, ctx, info) {
+        if (!ctx.request.userId) {
+            throw new Error('You must be logged in to do that.')
+        }
         const item = await ctx.db.mutation.createItem({
             data: {
-                ...args
+                user: {
+                    connect: {
+                        id: ctx.request.userId,
+                    },
+                },
+                ...args,
             }
         }, info)
 
@@ -26,7 +35,13 @@ const Mutations = {
     },
     async deleteItem (parent, args, ctx, info) {
         const where = { id: args.id };
-        const item = await ctx.db.query.item({ where }, `{id title }`);
+        const item = await ctx.db.query.item({ where }, `{id title user { id }}`);
+        const requiredPermissionsToDelete = ['ADMIN', 'ITEMDELETE'];
+        const ownsItem = item.user.id === ctx.request.userId;
+        const hasPermission = ctx.request.user.permissions.some(permission => requiredPermissionsToDelete.includes(permission));
+        if ( !ownsItem && !hasPermission ) {
+            throw new Error("You don't have the proper permissions to delete this item.")
+        };
         return ctx.db.mutation.deleteItem({ where }, info);
     },
     async signup (parent, args, ctx, info) {
@@ -118,6 +133,73 @@ const Mutations = {
         });
         return updatedUser;
     },
+    async updatePermissions (parent, args, ctx, info) {
+        if (!ctx.request.userId) {
+            throw new Error('You must be logged in to do that.')
+        }
+        const currentUser = await ctx.db.query.user({
+            where: {
+                id: ctx.request.userId,
+            },
+        }, info);
+        hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+        return ctx.db.mutation.updateUser({
+            data: {
+                permissions: {
+                    set: args.permissions,
+                },
+            },
+            where: {
+                id: args.userId,
+            },
+        }, info);
+    },
+    async addToCart (parent, args, ctx, info) {
+        const { userId } = ctx.request
+        if (!userId) {
+            throw new Error('You must be logged in to do that.')
+        }
+        const [existingCartItem] = await ctx.db.query.cartItems({
+            where: {
+                user: { id: userId },
+                item: { id: args.id },
+            },
+        }, info);
+        if (existingCartItem) {
+            return ctx.db.mutation.updateCartItem({
+                where: { id: existingCartItem.id },
+                data: { quantity: existingCartItem.quantity + 1 },
+            });
+        }
+        return ctx.db.mutation.createCartItem({
+            data: {
+                user: {
+                    connect: { id: userId },
+                },
+                item: {
+                    connect: { id: args.id },
+                },
+            },
+        }, info)
+    },
+    async removeFromCart (parent, args, ctx, info) {
+        const cartItem = await ctx.db.query.cartItem({
+            where: {
+                id: args.id
+            },
+        }, 
+        `{ id, user { id }}`
+        );
+        if (!cartItem) throw new Error('That item was not found in the cart');
+        if (cartItem.user.id !== ctx.request.userId) {
+            throw new Error('You do not have permission to delete that item.')
+        }
+        console.log('>>> delete cart item INFO ', info)
+        console.log('>>> delete cart item CTX ', ctx)
+        return ctx.db.mutation.deleteCartItem({
+            where: { id: args.id},
+        }, info)
+    }
 };
 
 module.exports = Mutations;
